requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-dotenv

langchain
langchain-community
langchain-chroma
langchain-ollama

pypdf
python-jose[cryptography]
passlib
python-docx 
python-pptx

langchain-huggingface
sentence-transformers
torch
openai
